chore: Rename Modal component to LoginModal and update import references
